requirements.txt
crispy-bootstrap5==0.6
Django==4.1
django-braces==1.15.0
django-cleanup==6.0.0
django-crispy-forms==1.14.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
django-global-permissions==0.2.6
django-role-permissions==3.1.1